refactor: move some functions of RemoteDebugger to mixins
