+++ lib/mixins/events.js
@@ -1,0 +1,9 @@
+// event emitted publically
+const events = {
+  EVENT_PAGE_CHANGE: 'remote_debugger_page_change',
+  EVENT_FRAMES_DETACHED: 'remote_debugger_frames_detached',
+  EVENT_DISCONNECT: 'remote_debugger_disconnect',
+};
+
+
+export default events;
