+++ lib/mixins/index.js
@@ -1,0 +1,7 @@
+import connect from './connect';
+import events from './events';
+import execute from './execute';
+import navigate from './navigate';
+
+
+export default Object.assign({}, connect, events, execute, navigate);
