+++ lib/rpc/remote-messages.js
@@ -1,6 +1,8 @@
 import _ from 'lodash';
 
 
+const OBJECT_GROUP = 'console';
+
 class RemoteMessages {
   constructor (isTargetBased = false) {
     this.isTargetBased = isTargetBased;
@@ -110,26 +112,24 @@ class RemoteMessages {
         message: JSON.stringify({
           id,
           method,
-          params: {
-            ...params,
-            objectGroup: 'console',
+          params: Object.assign({
+            objectGroup: OBJECT_GROUP,
             includeCommandLineAPI: true,
             doNotPauseOnExceptionsAndMuteConsole: false,
             emulateUserGesture: false,
             generatePreview: false,
             saveResult: false,
-          }
+          }, params)
         }),
       };
     } else {
       realMethod = method;
-      realParams = {
-        ...params,
-        objectGroup: 'console',
+      realParams = Object.assign({
+        objectGroup: OBJECT_GROUP,
         includeCommandLineAPI: true,
         doNotPauseOnExceptionsAndMuteConsole: false,
         emulateUserGesture: false,
-      };
+      }, params);
     }
 
     const plist = {
