+++ lib/rpc/rpc-client.js
@@ -382,10 +382,9 @@ export default class RpcClient {
 
     // highlight and then un-highlight the webview
     for (const enabled of [true, false]) {
-      await this.send('indicateWebView', {
-        ...sendOpts,
+      await this.send('indicateWebView', Object.assign({
         enabled,
-      }, false);
+      }, sendOpts), false);
     }
 
     await this.send('setSenderKey', sendOpts);
@@ -441,10 +440,9 @@ export default class RpcClient {
     // go through the tasks to initialize
     await this.send('Page.getResourceTree', sendOpts, false);
     await this.send('Network.enable', sendOpts, false);
-    await this.send('Network.setResourceCachingDisabled', {
-      ...sendOpts,
+    await this.send('Network.setResourceCachingDisabled', Object.assign({
       disabled: false,
-    }, false);
+    }, sendOpts), false);
     await this.send('DOMStorage.enable', sendOpts, false);
     await this.send('Database.enable', sendOpts, false);
     await this.send('IndexedDB.enable', sendOpts, false);
@@ -455,31 +453,25 @@ export default class RpcClient {
     await this.send('ApplicationCache.enable', sendOpts, false);
     await this.send('ApplicationCache.getFramesWithManifests', sendOpts, false);
     await this.send('Timeline.setInstruments', sendOpts, false);
-    await this.send('Timeline.setAutoCaptureEnabled', {
-      ...sendOpts,
+    await this.send('Timeline.setAutoCaptureEnabled', Object.assign({
       enabled: false,
-    }, false);
+    }, sendOpts), false);
     await this.send('Debugger.enable', sendOpts, false);
-    await this.send('Debugger.setBreakpointsActive', {
-      ...sendOpts,
+    await this.send('Debugger.setBreakpointsActive', Object.assign({
       active: true,
-    }, false);
-    await this.send('Debugger.setPauseOnExceptions', {
-      ...sendOpts,
+    }, sendOpts), false);
+    await this.send('Debugger.setPauseOnExceptions', Object.assign({
       state: 'none',
-    }, false);
-    await this.send('Debugger.setPauseOnAssertions', {
-      ...sendOpts,
+    }, sendOpts), false);
+    await this.send('Debugger.setPauseOnAssertions', Object.assign({
       enabled: false,
-    }, false);
-    await this.send('Debugger.setAsyncStackTraceDepth', {
-      ...sendOpts,
+    }, sendOpts), false);
+    await this.send('Debugger.setAsyncStackTraceDepth', Object.assign({
       depth: 200,
-    }, false);
-    await this.send('Debugger.setPauseForInternalScripts', {
-      ...sendOpts,
+    }, sendOpts), false);
+    await this.send('Debugger.setPauseForInternalScripts', Object.assign({
       shouldPause: false,
-    }, false);
+    }, sendOpts), false);
 
     await this.send('LayerTree.enable', sendOpts, false);
     await this.send('Worker.enable', sendOpts, false);
@@ -488,11 +480,10 @@ export default class RpcClient {
     await this.send('DOM.getDocument', sendOpts, false);
     const loggingChannels = await this.send('Console.getLoggingChannels', sendOpts);
     for (const source of (loggingChannels.channels || []).map((entry) => entry.source)) {
-      await this.send('Console.setLoggingChannelLevel', {
-        ...sendOpts,
+      await this.send('Console.setLoggingChannelLevel', Object.assign({
         source,
         level: 'verbose',
-      }, false);
+      }, sendOpts), false);
     }
 
     await this.send('Inspector.initialized', sendOpts, false);
